+++ tests/framework/django_test.py
@@ -10,8 +10,7 @@ from soapfish.testutil import echo_service, framework
 try:
     import django
     from django.conf import settings
-except (ImportError, SyntaxError):
-    # XXX: SyntaxError caused by Django 1.8+ on Python 2.6
+except ImportError:
     django = None
 else:
     settings.configure(
@@ -27,11 +26,6 @@ else:
     from django.conf.urls import url
     from django.test import Client
 
-if not hasattr(unittest, 'skip'):
-    # XXX: Skipping tests not supported in Python 2.6
-    import unittest2 as unittest
-
-
 urlconf = namedtuple('urlconf', 'urlpatterns')
 
 
@@ -39,16 +33,12 @@ urlconf = namedtuple('urlconf', 'urlpatterns')
 class DjangoDispatchTest(framework.DispatchTestMixin, unittest.TestCase):
 
     def setUp(self):  # noqa
-        # XXX: Python 2.6 and unittest2 still call this method for skipped class.
-        if django is None:
-            self.skipTest('Django is not installed.')
-
         self.service = echo_service()
         settings.ROOT_URLCONF = urlconf(urlpatterns=(url(r'^ws/$', django_dispatcher(self.service)),))
         self.client = Client()
 
     def _prepare_extras(self, headers):
-        extras = dict(('HTTP_' + k.replace('-', '_').upper(), v) for k, v in headers.items())
+        extras = {'HTTP_' + k.replace('-', '_').upper(): v for k, v in headers.items()}
         extras.update(content_type=headers['content-type'])
         return extras