+++ tests/framework/flask_test.py
@@ -11,19 +11,11 @@ try:
 except ImportError:
     flask = None
 
-if not hasattr(unittest, 'skip'):
-    # XXX: Skipping tests not supported in Python 2.6
-    import unittest2 as unittest
-
 
 @unittest.skipIf(flask is None, 'Flask is not installed.')
 class FlaskDispatchTest(framework.DispatchTestMixin, unittest.TestCase):
 
     def setUp(self):  # noqa
-        # XXX: Python 2.6 and unittest2 still call this method for skipped class.
-        if flask is None:
-            self.skipTest('Flask is not installed.')
-
         self.service = echo_service()
         app = flask.Flask(__name__)
         app.add_url_rule('/ws/', 'ws', flask_dispatcher(self.service), methods=['GET', 'POST'])