+++ tests/generation/xsd_code_test.py
@@ -14,10 +14,6 @@ from pythonic_testcase import (
 from soapfish import utils, xsd, xsd2py, xsdspec
 from soapfish.testutil import generated_symbols
 
-if not hasattr(unittest, 'skip'):
-    # XXX: Skipping tests not supported in Python 2.6
-    import unittest2 as unittest
-
 
 class XSDCodeGenerationTest(PythonicTestCase):
 
@@ -46,7 +42,7 @@ class XSDCodeGenerationTest(PythonicTestCase):
         assert_contains('Name', symbols.keys())
         assert_contains('Job', symbols.keys())
 
-        assert_equals(set(['name', 'job']), list(schemas[0].elements))
+        assert_equals({'job', 'name'}, list(schemas[0].elements))
 
         Job = symbols['Job']
         Name = symbols['Name']
@@ -78,7 +74,7 @@ class XSDCodeGenerationTest(PythonicTestCase):
         assert_contains('Person', symbols.keys())
         assert_contains('Job', symbols.keys())
 
-        assert_equals(set(['person', 'job']), list(schemas[0].elements))
+        assert_equals({'job', 'person'}, list(schemas[0].elements))
 
         Job = symbols['Job']
         Person = symbols['Person']