+++ CHANGES.md
@@ -49,7 +49,7 @@ the `soapfish` fork.
   - Attempts to fix handling of remote vs local imports.
 - **Miscellaneous:**
   - Renamed `SoapboxRequest` and `SoapboxResponse` to `SOAPRequest` and `SOAPResponse` respectively.
-  - Support Python 2.6, 2.7 and 3.3 - 3.6, and Django 1.6 - 1.11.
+  - Support Python 2.7 and 3.3 - 3.6, and Django 1.6 - 1.11.
   - Improved testing against different versions of Python, Django & Flask.
   - Improved entry points for generation scripts - additional flags, etc.
   - Moved to using an external dependency for `iso8601`